+++ packages/stream_chat_v1/lib/app/api_data.dart
@@ -1,0 +1,3 @@
+const kStreamApiKey = 'STREAM_API_KEY';
+const kStreamUserId = 'STREAM_USER_ID';
+const kStreamToken = 'STREAM_TOKEN';
