+++ packages/stream_chat_v1/lib/app/app_config.dart
