+++ packages/stream_chat_v1/lib/app/init_data.dart
@@ -1,0 +1,9 @@
+import 'package:stream_chat_flutter/stream_chat_flutter.dart';
+import 'package:streaming_shared_preferences/streaming_shared_preferences.dart';
+
+class InitData {
+  InitData(this.client, this.preferences);
+
+  final StreamChatClient client;
+  final StreamingSharedPreferences preferences;
+}
