+++ packages/stream_chat_v1/lib/app/routes/routes.dart
