+++ packages/stream_chat_v1/lib/app/splash_screen.dart
