+++ packages/stream_chat_v1/lib/screens/channel_file_display_screen.dart
@@ -1,4 +1,4 @@
-import 'package:example/localizations.dart';
+import 'package:example/app/localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:video_player/video_player.dart';
