+++ packages/stream_chat_v1/lib/screens/channel_list_page.dart
@@ -1,16 +1,15 @@
 import 'dart:async';
 
-import 'package:example/localizations.dart';
-import 'package:example/routes/routes.dart';
-import 'package:example/user_mentions_page.dart';
+import 'package:example/app/localizations.dart';
+import 'package:example/app/routes/routes.dart';
+import 'package:example/screens/user_mentions_page.dart';
+import 'package:example/widgets/channel_list.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app_badger/flutter_app_badger.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:streaming_shared_preferences/streaming_shared_preferences.dart';
-
-import 'channel_list.dart';
 
 class ChannelListPage extends StatefulWidget {
   const ChannelListPage({
