+++ packages/stream_chat_v1/lib/screens/channel_page.dart
@@ -1,11 +1,8 @@
 import 'package:collection/collection.dart';
-import 'package:example/routes/routes.dart';
-import 'package:example/thread_page.dart';
+import 'package:example/app/routes/routes.dart';
+import 'package:example/screens/screens.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-
-import 'chat_info_screen.dart';
-import 'group_info_screen.dart';
 
 class ChannelPageArgs {
   final Channel? channel;
