+++ packages/stream_chat_v1/lib/screens/group_info_screen.dart
@@ -1,16 +1,11 @@
 import 'dart:async';
 
 import 'package:collection/collection.dart' show IterableExtension;
-import 'package:example/localizations.dart';
+import 'package:example/app/localizations.dart';
+import 'package:example/screens/screens.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-
-import 'channel_file_display_screen.dart';
-import 'channel_media_display_screen.dart';
-import 'channel_page.dart';
-import 'chat_info_screen.dart';
-import 'pinned_messages_screen.dart';
 
 class GroupInfoScreen extends StatefulWidget {
   final StreamMessageThemeData messageTheme;
