+++ packages/stream_chat_v1/lib/screens/home_page.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:example/channel_page.dart';
-import 'package:example/notifications_service.dart';
-import 'package:example/routes/app_routes.dart';
-import 'package:example/routes/routes.dart';
+import 'package:example/app/routes/app_routes.dart';
+import 'package:example/app/routes/routes.dart';
+import 'package:example/screens/channel_page.dart';
+import 'package:example/services/notifications_service.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
