+++ packages/stream_chat_v1/lib/screens/new_chat_screen.dart
@@ -1,12 +1,11 @@
 import 'dart:async';
 
-import 'package:example/localizations.dart';
+import 'package:example/app/localizations.dart';
+import 'package:example/app/routes/routes.dart';
+import 'package:example/screens/channel_page.dart';
+import 'package:example/widgets/chips_input_text_field.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-
-import 'channel_page.dart';
-import 'chips_input_text_field.dart';
-import 'routes/routes.dart';
 
 class NewChatScreen extends StatefulWidget {
   @override
