+++ packages/stream_chat_v1/lib/screens/user_mentions_page.dart
@@ -1,9 +1,8 @@
-import 'package:example/localizations.dart';
-import 'package:example/routes/routes.dart';
+import 'package:example/app/localizations.dart';
+import 'package:example/app/routes/routes.dart';
+import 'package:example/screens/channel_page.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-
-import 'channel_page.dart';
 
 class UserMentionsPage extends StatefulWidget {
   @override
