+++ packages/stream_chat_v1/lib/services/notifications_service.dart
@@ -1,6 +1,6 @@
-import 'package:example/channel_page.dart';
-import 'package:example/localizations.dart';
-import 'package:example/routes/routes.dart';
+import 'package:example/screens/channel_page.dart';
+import 'package:example/app/localizations.dart';
+import 'package:example/app/routes/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart'
     hide Message;
