+++ packages/stream_chat_v1/lib/widgets/chips_input_text_field.dart
@@ -1,4 +1,4 @@
-import 'package:example/localizations.dart';
+import 'package:example/app/localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
