+++ packages/stream_chat_v1/lib/widgets/stream_version.dart
@@ -1,4 +1,4 @@
-import 'package:example/localizations.dart';
+import 'package:example/app/localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:yaml/yaml.dart';
